Move geometry helpers into runtime module
